Move ShareAcquireMode to internal package.
